refactor: discovery 폴더 구조 변경 (component -> page)
