+++ src/app/(discovery)/_components/DiscoveryInput.tsx
