+++ src/app/(discovery)/_components/DiscoverySection.tsx
