+++ src/app/(discovery)/_components/PopularSearches.tsx
