+++ src/app/(discovery)/_components/RecentSearches.tsx
