+++ src/app/(discovery)/_components/index.ts
