+++ src/app/(discovery)/layout.tsx
